load more thread replies when there are more to load
We currently cap the thread depth to avoid making to
many recursive calls on the backend when decorating
the comment tree for the thread page.

This adds a "see more" button that will fetch the
subcomments under the last node in a reply thread
if it has more replies to load.